Move DropdownMenu to Comment, include ChevronUpIcon when expanding, set the selection and focus when editing comment
