+++ src/recogito-client-core/editor/widgets/comment/TextEntryField.jsx
@@ -20,8 +20,27 @@ export default class TextEntryField extends Component {
   }
 
   componentDidMount() {
-    if (this.props.focus && this.element.current)
-      this.element.current.focus({ preventScroll: true });
+    if (this.props.focus && this.element.current) {
+      this.setSelectionAndFocus()
+    }
+  }
+
+  componentDidUpdate(prevProps) {
+    if (this.props.editable !== prevProps.editable) {
+      this.setSelectionAndFocus()
+    }
+  }
+
+  setSelectionAndFocus() {
+    const {content, editable} = this.props
+    if (!this.element.current || !editable) {
+      return
+    }
+    if (content) {
+      // Set the selection to the end of content, not the beginning
+      this.element.current.setSelectionRange(content.length, content.length)
+    }
+    this.element.current.focus({ preventScroll: true });
   }
 
   render() {
@@ -33,7 +52,7 @@ export default class TextEntryField extends Component {
         placeholder={this.props.placeholder || i18n.t('Add a comment...')}
         disabled={!this.props.editable}
         onChange={this.props.onChange}
-        onKeyDown={this.onKeyDown} />
+        onKeyDown={this.onKeyDown}/>
     )
   }
